+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## next
+
+- Fixed nested ternary operator precendence
+
 ## 1.0.0-alpha.5 (November 26, 2018)
 
 - Added computed property names support in object literals