+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fixed nested ternary operator precendence
 - Added destuction for variables when used on object literal with no value (i.e. `{ $foo, bar: 1}` the same as `{ foo: $foo, bar: 1 }`)
 - Changed `in` and `not in` operators to propertly work with an object and a string on right side
+- Added support for a function as a parameter for `pick()` method
 
 ## 1.0.0-alpha.5 (November 26, 2018)